refactor(store): split the recorder's IO shell and counter math out
tools/record.py carried the record contract, the git and forge shell,
the preference-counter math and the five verbs in 994 lines. The shell
(store checkout, git, PR call) becomes record_store.py and the
counter math with the PR body it explains becomes record_confirm.py;
record.py keeps the docstring that is the CLI help, the decision-shaped
record policy, the verbs and main.

store_root() stays with the verbs on purpose: it answers "which
checkout am I in" from its own __file__, and a recorder relocated into
a store must get that answer from the entry it was run as.

The fixtures that build a throwaway store copy every tools/*.py rather
than record.py alone — the recorder is three modules now, and the old
single-file copy only worked because an earlier import had left the
real directory on sys.path.

ADVANCES pandoscope/agentic-engineering-template#239.
